tests/wpt/tests/websockets/interfaces/WebSocket/readyState/004.html

<!doctype html>
<title>WebSockets: defineProperty getter for readyState</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.sub.js></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
test(function() {
    Object.defineProperty(WebSocket.prototype, 'readyState', {
        get: function() { return 'foo'; }
    });
    var ws = new WebSocket('ws://example.invalid/');
    assert_equals(ws.readyState, 'foo');
});
</script>
